consolidate style files
