+++ client/src/admin/Library/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import cn from 'classnames';
 
-import { flex, p1, spacedEvenly, flex1, column } from '../styles';
+import { flex, p1, spacedEvenly, flex1, column } from '../../styles';
 import { getLibrary } from '../api';
 import AudioMetadata from './AudioMetadata';
 import AudioUpload from './AudioUpload';
